moved tri from math to shape
